Highlights da quali

+++ src/asymmetricdemonsfunction.cxx
@@ -5,12 +5,12 @@ VectorField AsymmetricDemonsFunction::newDeltaField(VectorField gradients) {
 	for(int x = 0; x < dimensions[0]; x++)
 		for(int y = 0; y < dimensions[1]; y++)
 			for(int z = 0; z < dimensions[2]; z++) {
-				std::vector<double> sGradVec = gradients.getVectorAt(x, y, z);
-				double deformedValue = getDeformedImageValueAt(x, y, z);
-				double diff = deformedValue - staticImage_.getPixelAt(x,y,z);
-				double denominator = ((diff*diff)/spacing) + (sGradVec[0])*(sGradVec[0]) + (sGradVec[1])*(sGradVec[1]) + (sGradVec[2])*(sGradVec[2]);
+				std::vector<float> sGradVec = gradients.getVectorAt(x, y, z);
+				float deformedValue = getDeformedImageValueAt(x, y, z);
+				float diff = deformedValue - staticImage_.getPixelAt(x,y,z);
+				float denominator = ((diff*diff)/spacing) + (sGradVec[0])*(sGradVec[0]) + (sGradVec[1])*(sGradVec[1]) + (sGradVec[2])*(sGradVec[2]);
 				if (denominator > 0.0) {
-					std::vector<double> deltaVector(3, 0.0);
+					std::vector<float> deltaVector(3, 0.0);
 					deltaVector[0] = (sGradVec[0])*diff/denominator;
 					deltaVector[1] = (sGradVec[1])*diff/denominator;
 					deltaVector[2] = (sGradVec[2])*diff/denominator;
