+++ src/vectorfield.cxx
@@ -18,23 +18,23 @@ VectorField::VectorField(VectorField3D vectorField) {
     this->vectorField = vectorField;
 }
 
-VectorField::VectorField(std::vector<int> dimensions, double defaultValue) {
+VectorField::VectorField(std::vector<int> dimensions, float defaultValue) {
 	this->dimensions = dimensions;
 	vectorField = createVectorField3D(dimensions, defaultValue);
 }
 
-VectorField::VectorField3D VectorField::createVectorField3D(std::vector<int> dimensions, double defaultValue) {
+VectorField::VectorField3D VectorField::createVectorField3D(std::vector<int> dimensions, float defaultValue) {
 	using std::vector;
-	return VectorField3D(dimensions[0], vector<vector<vector<double> > >(dimensions[1], vector<vector<double> >(dimensions[2], vector<double>(3, defaultValue))));
+	return VectorField3D(dimensions[0], vector<vector<vector<float> > >(dimensions[1], vector<vector<float> >(dimensions[2], vector<float>(3, defaultValue))));
 }
 
-std::vector<double> VectorField::getVectorAt(int x, int y, int z) {
+std::vector<float> VectorField::getVectorAt(int x, int y, int z) {
 	if ((x < 0 || x > dimensions[0]-1) || (y < 0 || y > dimensions[1]-1) || (z < 0 || z > dimensions[2]-1)) 
-		return std::vector<double>(3, 0.0);
+		return std::vector<float>(3, 0.0);
 	return vectorField[x][y][z];
 }
 
-void VectorField::updateVector(Demon demon, std::vector<double> newValues) {
+void VectorField::updateVector(Demon demon, std::vector<float> newValues) {
 	std::vector<int> position = demon.getPosition();
 	std::vector<int> volumeOfInfluence = demon.getVolumeOfInfluence();
 	for (int x = position[0]-volumeOfInfluence[0]; x < position[0]+volumeOfInfluence[0]; x++)
@@ -50,11 +50,11 @@ void VectorField::updateVector(Demon demon, std::vector<double> newValues) {
 			}
 }
 
-void VectorField::updateVector(int x, int y, int z, std::vector<double> newValues) {
+void VectorField::updateVector(int x, int y, int z, std::vector<float> newValues) {
 	vectorField[x][y][z] = newValues;
 }
 
-// double VectorField::vectorNorm(std::vector<double> v) {
+// float VectorField::vectorNorm(std::vector<float> v) {
 // 	return sqrt(pow(v[0],2)+pow(v[1],2));
 // }
 
@@ -62,8 +62,8 @@ void VectorField::updateVector(int x, int y, int z, std::vector<double> newValue
 // 	VectorField normalized(rows_, cols_);
 // 	for(int row = 0; row < rows_; row++) {
 // 	    for(int col = 0; col < cols_; col++) {
-// 	    	std::vector<double> vector = vectorField[row][col];
-// 	    	double normalizedRow = 0.0, normalizedCol = 0.0;
+// 	    	std::vector<float> vector = vectorField[row][col];
+// 	    	float normalizedRow = 0.0, normalizedCol = 0.0;
 // 	    	if ((vector[0]*vector[1]) != 0.0) {
 // 	    		normalizedRow = vector[0]/vectorNorm(vector);
 // 	    		normalizedCol = vector[1]/vectorNorm(vector);
@@ -74,15 +74,15 @@ void VectorField::updateVector(int x, int y, int z, std::vector<double> newValue
 // 	return normalized;
 // }
 
-std::vector<double> VectorField::generateGaussianFilter2D(int kernelSize, double deviation) {
-	std::vector<double> gaussianFilter(kernelSize, 0.0);
-	double sum = 0.0;
+std::vector<float> VectorField::generateGaussianFilter2D(int kernelSize, float deviation) {
+	std::vector<float> gaussianFilter(kernelSize, 0.0);
+	float sum = 0.0;
 	for (int i = 0; i < kernelSize; i++) {
-		double firstTerm = (i-(kernelSize-1)/2.0);
-		double squaredFirstTerm = std::pow(firstTerm,2);
-		double secondTerm = deviation;
-		double squaredSecondTerm = 2.0*std::pow(secondTerm,2);
-		double finalTerm = -1.0*squaredFirstTerm/squaredSecondTerm;
+		float firstTerm = (i-(kernelSize-1)/2.0);
+		float squaredFirstTerm = std::pow(firstTerm,2);
+		float secondTerm = deviation;
+		float squaredSecondTerm = 2.0*std::pow(secondTerm,2);
+		float finalTerm = -1.0*squaredFirstTerm/squaredSecondTerm;
 		gaussianFilter[i] = std::exp(finalTerm);
 		sum += gaussianFilter[i];
 	}
@@ -91,23 +91,23 @@ std::vector<double> VectorField::generateGaussianFilter2D(int kernelSize, double
 	return gaussianFilter;
 }
 
-bool isZero(std::vector<double> vector) {
+bool isZero(std::vector<float> vector) {
 	if ((vector[0] <= EPSILON) && (vector[1] <= EPSILON) && (vector[2] <= EPSILON)) return true;
 	return false;
 }
 
-void VectorField::applyGaussianFilter(int kernelSize, double deviation) {
+void VectorField::applyGaussianFilter(int kernelSize, float deviation) {
 	Profiler profiler("applyGaussianFilter");
-	std::vector<double> gaussianKernel = generateGaussianFilter2D(kernelSize, deviation);
+	std::vector<float> gaussianKernel = generateGaussianFilter2D(kernelSize, deviation);
 	VectorField3D auxVectorField = vectorField;
 	for(int x = 0; x < dimensions[0]; x++)
 		for(int y = 0; y < dimensions[1]; y++)
 			for(int z = 0; z < dimensions[2]; z++) {
-				std::vector<double> vector = vectorField[x][y][z];
+				std::vector<float> vector = vectorField[x][y][z];
 				if (isZero(vector)) continue;
-				std::vector<double> newValues(3, 0.0);
+				std::vector<float> newValues(3, 0.0);
 				for (int i = -1; i <= 1; i++) {
-					std::vector<double> auxVector = getVectorAt(x+i,y,z);
+					std::vector<float> auxVector = getVectorAt(x+i,y,z);
 					newValues[0] += auxVector[0]*gaussianKernel[i+1];
 					newValues[1] += auxVector[1]*gaussianKernel[i+1];
 					newValues[2] += auxVector[2]*gaussianKernel[i+1];
@@ -118,11 +118,11 @@ void VectorField::applyGaussianFilter(int kernelSize, double deviation) {
 	for(int x = 0; x < dimensions[0]; x++)
 		for(int y = 0; y < dimensions[1]; y++)
 			for(int z = 0; z < dimensions[2]; z++) {
-				std::vector<double> vector = vectorField[x][y][z];
+				std::vector<float> vector = vectorField[x][y][z];
 				if (isZero(vector)) continue;
-				std::vector<double> newValues(3, 0.0);
+				std::vector<float> newValues(3, 0.0);
 				for (int i = -1; i <= 1; i++) {
-					std::vector<double> auxVector = getVectorAt(x,y+i,z);
+					std::vector<float> auxVector = getVectorAt(x,y+i,z);
 					newValues[0] += auxVector[0]*gaussianKernel[i+1];
 					newValues[1] += auxVector[1]*gaussianKernel[i+1];
 					newValues[2] += auxVector[2]*gaussianKernel[i+1];
@@ -133,11 +133,11 @@ void VectorField::applyGaussianFilter(int kernelSize, double deviation) {
 	for(int x = 0; x < dimensions[0]; x++)
 		for(int y = 0; y < dimensions[1]; y++)
 			for(int z = 0; z < dimensions[2]; z++) {
-				std::vector<double> vector = vectorField[x][y][z];
+				std::vector<float> vector = vectorField[x][y][z];
 				if (isZero(vector)) continue;
-				std::vector<double> newValues(3, 0.0);
+				std::vector<float> newValues(3, 0.0);
 				for (int i = -1; i <= 1; i++) {
-					std::vector<double> auxVector = getVectorAt(x,y,z+i);
+					std::vector<float> auxVector = getVectorAt(x,y,z+i);
 					newValues[0] += auxVector[0]*gaussianKernel[i+1];
 					newValues[1] += auxVector[1]*gaussianKernel[i+1];
 					newValues[2] += auxVector[2]*gaussianKernel[i+1];
@@ -158,11 +158,11 @@ void VectorField::add(VectorField adding) {
 			}
 }
 
-// double VectorField::sumOfAbs() {
-// 	double total = 0.0;
+// float VectorField::sumOfAbs() {
+// 	float total = 0.0;
 // 	for(int row = 0; row < rows_; row++) {
 // 		for(int col = 0; col < cols_; col++) {
-// 			total += std::abs(vectorCol_.at<double>(row, col)) + std::abs(vectorRow_.at<double>(row, col));
+// 			total += std::abs(vectorCol_.at<float>(row, col)) + std::abs(vectorRow_.at<float>(row, col));
 // 		}
 // 	}
 // 	return total;
@@ -174,7 +174,7 @@ void VectorField::printAround(int x, int y, int z) {
 		for (int aroundY = y-1; aroundY <= y+1; aroundY++) {
 			std::cout << "\n";
 			for (int aroundZ = z-1; aroundZ <= z+1; aroundZ++) {
-				std::vector<double> macaco = getVectorAt(aroundX,aroundY,aroundZ);
+				std::vector<float> macaco = getVectorAt(aroundX,aroundY,aroundZ);
 				std::cout << "(" << macaco[0] << ", " << macaco[1] << ", " << macaco[2] << ")\t";
 			}
 		}
@@ -196,14 +196,14 @@ void VectorField::printAround(int x, int y, int z) {
 // // 	imwrite(flCol.c_str(), abs_grad_col, compression_params);
 // // }
 
-// // std::vector<double> VectorField::getInfos() {
-// // 	std::multiset<double> magnitudes;
+// // std::vector<float> VectorField::getInfos() {
+// // 	std::multiset<float> magnitudes;
 // // 	int size = (rows_*cols_);
-// // 	double max= 0.0, min = 0.0, mean = 0.0, median = 0.0, deviation = 0.0;
+// // 	float max= 0.0, min = 0.0, mean = 0.0, median = 0.0, deviation = 0.0;
 // 	// for(int row = 0; row < rows_; row++) {
 // 	//     for(int col = 0; col < cols_; col++) {
-// 	//     	std::vector<double> vector = getVectorAt(row, col);
-//  //    		double mag = std::sqrt(vector[0]*vector[0] + vector[1]*vector[1]);
+// 	//     	std::vector<float> vector = getVectorAt(row, col);
+//  //    		float mag = std::sqrt(vector[0]*vector[0] + vector[1]*vector[1]);
 //  //    		if (max < mag) max = mag;
 //  //    		if (min > mag) min = mag;
 //  //    		mean += mag;
@@ -212,7 +212,7 @@ void VectorField::printAround(int x, int y, int z) {
 // 	// }
 // // 	mean /= size;
 // // 	int count = 1;
-// // 	std::multiset<double>::iterator it;
+// // 	std::multiset<float>::iterator it;
 // // 	for (it=magnitudes.begin(); it!=magnitudes.end(); ++it) {
 // //     	deviation += std::pow((*it - mean),2);
 // //     	if (count == size/2) median = *it;
@@ -220,7 +220,7 @@ void VectorField::printAround(int x, int y, int z) {
 // // 	}
 // // 	deviation /= size;
 // // 	deviation = std::sqrt(deviation);
-// // 	std::vector<double> results;
+// // 	std::vector<float> results;
 // // 	results.push_back(min);
 // // 	results.push_back(max);
 // // 	results.push_back(median);
@@ -234,7 +234,7 @@ void VectorField::printAround(int x, int y, int z) {
 // // 	if (iteration <= 1) myfile.open(filename);
 // // 	else myfile.open(filename, std::ios_base::app);
 // // 	myfile << "Iteration " << iteration << "\n";
-// // 	std::vector<double> results = getInfos();
+// // 	std::vector<float> results = getInfos();
 // // 	myfile << "Min = " << results[0] << " Max = \t" << results[1] << " Median = \t" << results[2] << " Mean = \t" << results[3] << " Standard Deviaon = \t" << results[4] << "\n";
 // // 	myfile.close();
 // // }
@@ -242,17 +242,17 @@ void VectorField::printAround(int x, int y, int z) {
 // // void VectorField::printField(std::string filename) {
 // // 	std::ofstream myfile;
 // // 	myfile.open(filename);
-// // 	double minValCol, maxValCol;
-// // 	double minValRow, maxValRow;
+// // 	float minValCol, maxValCol;
+// // 	float minValRow, maxValRow;
 // // 	minMaxLoc(vectorCol_, &minValCol, &maxValCol);
 // // 	minMaxLoc(vectorRow_, &minValRow, &maxValRow);
 // // 	for(int row = 0; row < rows_; row++) {
 // // 	    for(int col = 0; col < cols_; col++) {
-// // 	    	std::vector<double> vector = getVectorAt(row, col);
-// //     		double redCol = 255*(vector[1]-minValCol)/(maxValCol-minValCol);
-// // 			double blueCol = 255*(maxValCol-vector[1])/(maxValCol-minValCol);
-// // 			double redRow = 255*(vector[0]-minValRow)/(maxValRow-minValRow);
-// // 			double blueRow = 255*(maxValRow-vector[0])/(maxValRow-minValRow);
+// // 	    	std::vector<float> vector = getVectorAt(row, col);
+// //     		float redCol = 255*(vector[1]-minValCol)/(maxValCol-minValCol);
+// // 			float blueCol = 255*(maxValCol-vector[1])/(maxValCol-minValCol);
+// // 			float redRow = 255*(vector[0]-minValRow)/(maxValRow-minValRow);
+// // 			float blueRow = 255*(maxValRow-vector[0])/(maxValRow-minValRow);
 // // 			int red = (redCol + redRow)/2;
 // // 			int blue = (blueCol + blueRow)/2;
 // // 			myfile << col << " " << (vectorCol_.rows - row) << " " << vector[1] << " " << vector[0] << " " <<  red << " " << 0 << " " << blue << "\n";
